Downloads/Tsunade/Tsunade/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.20
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7
redis==5.0.1
cryptography==40.0.2
fastapi-users[sqlalchemy]==12.1.3
httpx-oauth==0.16.1
pyjwt==2.8.0
email-validator==2.1.0
argon2-cffi==23.1.0
slowapi==0.1.9
httpx==0.25.2
openai==1.3.7
pillow==10.1.0
requests==2.31.0
numpy==1.24.4
gunicorn==21.2.0
motor==3.3.2
pytesseract==0.3.10
python-jose[cryptography]==3.3.0
sendgrid==6.10.0
firebase-admin==6.2.0
fuzzywuzzy==0.18.0
python-levenshtein==0.21.1
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
aiosqlite==0.19.0
asyncpg==0.29.0
sentry-sdk[fastapi]==1.38.0